backend/sentiment/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
pydantic==2.9.2
prometheus-client==0.20.0
python-dotenv==1.0.1
PyYAML==6.0.2
numpy==2.1.1
